Lanes start with one vehicle, add vehicles over 3 second interval, and reset upon collision with player

+++ FroggerStarter/Constants/LaneSettings.cs
@@ -20,7 +20,7 @@
         /// <summary>
         ///     The lane length
         /// </summary>
-        public static int LaneLength = 650;
+        public static double LaneLength = 650;
 
         #endregion
     }
